+++ MidaiEsfe/MidaiEsfe.Aplicacion.EntidadesDeNegocio/Notas.cs
@@ -8,12 +8,12 @@ namespace MidaiEsfe.Aplicacion.EntidadesDeNegocio
 {
     public class Notas
     {
-        public byte Id { get; set; }
+        public Int64 Id { get; set; }
 
-        public byte IdEvaluacion { get; set; }
+        public Int64 IdEvaluacion { get; set; }
 
-        public byte IdAsignacionDeModulo { get; set; }
+        public Int64 IdAsignacionDeModulo { get; set; }
 
-        public byte Nota { get; set; }
+        public Int64 Nota { get; set; }
     }
 }
